de-project-final/requirements.txt
pandas
vertica_python
apache-airflow-providers-vertica
apache-airflow-providers-common-sql

